fix seperation
