+++ tauri-build/src/lib.rs
@@ -5,18 +5,68 @@ extern crate serde_json;
 #[macro_use]
 extern crate lazy_static;
 
-extern crate tauri;
-
-use std::env;
-use std::io::Write;
-
-pub mod config;
 mod api;
-mod salt;
+pub mod config;
 pub mod event;
+pub mod server;
+
+mod file_system;
+mod salt;
 
 #[cfg(feature = "embedded-server")]
 mod tcp;
 
 mod app;
+
+use std::process::Stdio;
+
+use threadpool::ThreadPool;
+
 pub use app::*;
+use web_view::*;
+
+thread_local!(static POOL: ThreadPool = ThreadPool::new(4));
+
+pub fn spawn<F: FnOnce() -> () + Send + 'static>(what: F) {
+  POOL.with(|thread| {
+    thread.execute(move || {
+      what();
+    });
+  });
+}
+
+pub fn execute_promise<T: 'static, F: FnOnce() -> Result<String, String> + Send + 'static>(
+  webview: &mut WebView<'_, T>,
+  what: F,
+  callback: String,
+  error: String,
+) {
+  let handle = webview.handle();
+  POOL.with(|thread| {
+    thread.execute(move || {
+      let callback_string = tauri::rpc::format_callback_result(what(), callback, error);
+      handle
+        .dispatch(move |_webview| _webview.eval(callback_string.as_str()))
+        .unwrap()
+    });
+  });
+}
+
+pub fn call<T: 'static>(
+  webview: &mut WebView<'_, T>,
+  command: String,
+  args: Vec<String>,
+  callback: String,
+  error: String,
+) {
+  execute_promise(
+    webview,
+    || {
+      tauri::command::get_output(command, args, Stdio::piped())
+        .map_err(|err| format!("`{}`", err))
+        .map(|output| format!("`{}`", output))
+    },
+    callback,
+    error,
+  );
+}
