+++ tauri/Cargo.toml
@@ -10,13 +10,9 @@ edition = "2018"
 exclude = ["test/fixture/**"]
 
 [dependencies]
-web-view = "0.5.4"
-serde_json = "1.0.44"
-serde = "1.0"
-serde_derive = "1.0"
+serde = { version = "1.0", features = ["derive"] } 
 dirs = "2.0.2"
 ignore = "0.4.10"
-threadpool = "1.7"
 rand = "0.7"
 zip = "0.5.3"
 tempdir = "0.3"
@@ -26,5 +22,5 @@ either = "1.5.3"
 tar = "0.4"
 flate2 = "1"
 sysinfo = "0.10"
-tiny_http = "0.6"
+
